82929 updated FM36 model

+++ src/FM36/ESFA.DC.ILR.FundingService.FM36.FundingOutput.Model/Output/PriceEpisodeValues.cs
@@ -70,7 +70,7 @@ namespace ESFA.DC.ILR.FundingService.FM36.FundingOutput.Model.Output
 
         public decimal? PriceEpisodeExpectedTotalMonthlyValue { get; set; }
 
-        public long? PriceEpisodeAimSeqNumber { get; set; }
+        public int? PriceEpisodeAimSeqNumber { get; set; }
 
         public decimal? PriceEpisodeFirstDisadvantagePayment { get; set; }
